Create Bot-Activity-Idiot-Light.pine
Indicates a bot activity.
A red dot at the top of a candle means that conditions of the price action may be bullish bot manipulation. 

Conversely, a green dot underneath a candle indicates that a there's a possibility that bots are running out of steam and a reversal may be possible.

Lastly, a cyan or aqua triangle indicates that a bots may have become bullish and assisting in the trend upwards.

This script has been well tested over several months.  When you see a red dot, there's a pretty good chance of a down turn.  That's not financial advice, period.  

TL;DR;  Green is good. Red is bad.  Clusters of green dots and aqua triangles can be a good thing.

happy trading and good luck!